@@ -16,13 +16,13 @@
         "require": {
     		"php": ">=7.2",
 			"slim/slim": "^4.5",
-			"slim/twig-view": "^2.0",
-			"respect/validation": "~1.0",
+			"slim/twig-view": "^3.0",
+			"respect/validation": "^2.2",
 			"illuminate/database": "^8.22.1",
-		    "monolog/monolog": "^1.17",
+		    "monolog/monolog": "^2.2",
 			"slim/flash": "^0.4.0",
 			"illuminate/translation": "^8.23.1",
-			"symfony/event-dispatcher": "^4.3",
+			"symfony/event-dispatcher": "^5.0",
 		    "slim/psr7": "^0.5.0",
 		    "php-di/php-di": "^6.3",
 			"phpmailer/phpmailer": "^6.1",
