@@ -36,11 +36,11 @@ class FileUpdateController extends Controller
             }
             
             
-            if ($v['md5'] == 'new' && $v['updated'] == 0) {           
+            if ($v['md5'] == 'new' && $v['updated'] == 0) {
                 $pathHandler = explode('/', substr($v['path'], 1));
                 $pathCheck = MAIN_DIR.'/';
                
-                foreach ($pathHandler as $newPath) {    
+                foreach ($pathHandler as $newPath) {
                     $pathCheck .= $newPath;
                     
                     if (!is_dir($pathCheck)) {
