@@ -6,11 +6,11 @@ namespace Application\Core\Modules\Validation\Exceptions;
 
 use Respect\Validation\Exceptions\ValidationException;
 
-class EmailAvailbleException extends ValidationException
+final class EmailAvailbleException extends ValidationException
 {
-    public static $defaultTemplates = [
+    protected $defaultTemplates = [
         self::MODE_DEFAULT => [
-            self::STANDARD => 'Ten email jest już w bazie.',
-        ],
+            self::STANDARD => 'This email is in database',
+        ]
     ];
 }
