@@ -7,9 +7,9 @@ namespace Application\Core\Modules\Validation\Rules;
 use Application\Models\UserModel;
 use Respect\Validation\Rules\AbstractRule;
 
-class EmailAvailble extends AbstractRule
+final class EmailAvailble extends AbstractRule
 {
-    public function validate($input)
+    public function validate($input) : bool
     {
         return UserModel::where('email', $input)->count() === 0;
     }
