@@ -12,6 +12,6 @@ class SecretModel extends Model
     protected $fillable =
         [
             'user_id',
-            'secret'			
+            'secret'
         ];
 }
