@@ -13,26 +13,26 @@ class MessageController extends Controller
 {
     public function index($request, $response, $arg)
     {
-	    if(isset($_SESSION['user'])){
-			$this->view->getEnvironment()->addGlobal('unread', self::countUnreadMessages());
-            $this->view->getEnvironment()->addGlobal('folder', $arg['folder']);  
+        if (isset($_SESSION['user'])) {
+            $this->view->getEnvironment()->addGlobal('unread', self::countUnreadMessages());
+            $this->view->getEnvironment()->addGlobal('folder', $arg['folder']);
             $this->view->getEnvironment()->addGlobal('messages', self::getMessages($arg['folder']));
-            if($request->getParsedBody() !== null){
+            if ($request->getParsedBody() !== null) {
                 $body = $request->getParsedBody();
                 $body['body'] = str_replace('<p>', '<p>| ', $body['body']);
                 $this->view->getEnvironment()->addGlobal('reply', $body);
             }
-	    }	
+        }
 
-		return $this->view->render($response, 'mailbox.twig');
-	}
+        return $this->view->render($response, 'mailbox.twig');
+    }
 
-	public function sendMessage($request, $response)
-	{
-		$body = $request->getParsedBody();
-		if(isset($body['topic']) && isset($body['body'])){
+    public function sendMessage($request, $response)
+    {
+        $body = $request->getParsedBody();
+        if (isset($body['topic']) && isset($body['body'])) {
             $recipient = UserModel::where('username', $body['recipient'])->first();
-            if ($recipient === null ) {
+            if ($recipient === null) {
                 $this->flash->addMessage('danger', 'recipient does not exists');
                 return  $response
                 ->withHeader('Location', $this->router->urlFor('mailbox', ['folder' => 'outbox']))
@@ -45,11 +45,11 @@ class MessageController extends Controller
                 'topic' => $this->purifier->purify($body['topic']),
                 'body' => $this->purifier->purify($body['body'])
             ]);
-        }else{
+        } else {
             $this->flash->addMessage('danger', 'You cant send massage without topic or content!');
         }
 
-        if(isset($message)){
+        if (isset($message)) {
             MailboxModel::create([
                 'message_id' => $message->id,
                 'mailbox' => 'inbox',
@@ -65,24 +65,24 @@ class MessageController extends Controller
             ]);
             
             $this->flash->addMessage('success', 'Message send!');
-        }else{
+        } else {
             $message->delete();
             $this->flash->addMessage('danger', 'Something went wrong!');
         }
         
-		return  $response
+        return  $response
             ->withHeader('Location', $this->router->urlFor('mailbox', ['folder' => 'outbox']))
             ->withStatus(302);
-	}
-	
-	public function getMessage($request, $response, $arg)
-	{
-		$message = MessageModel::where('id', $arg['id'])->first();
+    }
+    
+    public function getMessage($request, $response, $arg)
+    {
+        $message = MessageModel::where('id', $arg['id'])->first();
        
-		if($message->sender_id === $_SESSION['user'] || $message->recipient_id === $_SESSION['user']){
-            if($message->recipient_id === $_SESSION['user']){
+        if ($message->sender_id === $_SESSION['user'] || $message->recipient_id === $_SESSION['user']) {
+            if ($message->recipient_id === $_SESSION['user']) {
                 $unread = MailboxModel::where([
-                    ['user_id', $_SESSION['user']], 
+                    ['user_id', $_SESSION['user']],
                     ['message_id', $arg['id']]
                 ])->first();
                 $this->cache->setName('unread-messages');
@@ -96,11 +96,11 @@ class MessageController extends Controller
             ];
             $this->view->getEnvironment()->addGlobal('users', $users);
             $this->view->getEnvironment()->addGlobal('message', $message);
-		}
+        }
         $this->view->getEnvironment()->addGlobal('box', $arg['folder']);
         $this->view->getEnvironment()->addGlobal('folder', 'message');
         return $this->view->render($response, 'mailbox.twig');
-	}
+    }
     
     public function deleteMessage($request, $response)
     {
@@ -116,7 +116,7 @@ class MessageController extends Controller
     {
         $body = $request->getParsedBody();
         $message = MailboxModel::where([
-            ['user_id', $_SESSION['user']], 
+            ['user_id', $_SESSION['user']],
             ['message_id', $body['id']]
         ])->first();
         $message->mailbox = $body['newFolder'];
@@ -130,15 +130,15 @@ class MessageController extends Controller
     public function countUnreadMessages(): int
     {
         $this->cache->setName('unread-messages');
-        if(isset($_SESSION['user'])){
-                if(!$unread = $this->cache->receive('user-'.$_SESSION['user'])){
-                    $unread =  MailboxModel::where([
-                        ['user_id', $_SESSION['user']], 
+        if (isset($_SESSION['user'])) {
+            if (!$unread = $this->cache->receive('user-'.$_SESSION['user'])) {
+                $unread =  MailboxModel::where([
+                        ['user_id', $_SESSION['user']],
                         ['unread', 1]
-                    ])->count(); 
-                    $this->cache->store('user-'.$_SESSION['user'], $unread, $this->settings['cache']['cache_time']);
-                }
-        }else{
+                    ])->count();
+                $this->cache->store('user-'.$_SESSION['user'], $unread, $this->settings['cache']['cache_time']);
+            }
+        } else {
             $unread = 0;
         }
         
@@ -148,9 +148,9 @@ class MessageController extends Controller
     public function getMessages(string $folder): array
     {
         $messages = [];
-        if(isset($_SESSION['user'])){
+        if (isset($_SESSION['user'])) {
             $messages = MailboxModel::where([
-                ['user_id', $_SESSION['user']], 
+                ['user_id', $_SESSION['user']],
                 ['mailbox', $folder]
             ])->orderBy('message.created_at', 'DESC')
             ->leftJoin('message', 'mailbox.message_id', 'message.id')
@@ -161,4 +161,3 @@ class MessageController extends Controller
         return $messages;
     }
 };
-    
