@@ -1,11 +1,11 @@
 <?php
-		return [
-			'driver' => 'mysql',
-			'host' => 'mysql.ct8.pl',
-			'database' => '',
-			'username' => '',
-			'password' => '',
-			'charset' => 'utf8',
-			'collation' => 'utf8_general_ci',
-			'prefix' => 'brd_'
-		];
+        return [
+            'driver' => 'mysql',
+            'host' => 'mysql.ct8.pl',
+            'database' => '',
+            'username' => '',
+            'password' => '',
+            'charset' => 'utf8',
+            'collation' => 'utf8_general_ci',
+            'prefix' => 'brd_'
+        ];
