@@ -79,7 +79,7 @@ return [
     'back' => 'Powrót',
     'inside-top' => 'Wewnętrzna kolumna góra',
     'inside-bottom' => 'Wewnętrzna kolumna dół',
-    'add group' => 'Dodaj grupę', 
+    'add group' => 'Dodaj grupę',
     'Are you sure you want remove permanently this group' => 'Czy na pewno chcesz trwale usunać tą grupę',
     'delete group' => 'Usuń grupę',
     'is active' => 'aktywna',
@@ -107,6 +107,7 @@ return [
     'plugin info' => 'Informacje o pluginie',
     'confirm_reg' => 'Potwierdzenie rejestracji przez email',
     'default_group' => 'Domyslna grupa nowych użytkowników',
+    'menu.manage' => 'Zarządzanie linkiem',
     
     
     'update' => 'Aktualizacja',
@@ -121,18 +122,37 @@ return [
     'deactive' => 'Wyłącz',
     'activate' => 'Włącz',
     'uninstall' => 'Odinstaluj',
+    'author' => 'Autor',
+    'author website' => 'Strona',
+    'description' => 'Opis',
+    'plugins.ext' => 'Zarządzanie pluginem',
     
-    'admin.boards' => 'Forum',
-    'admin.get.settings' => 'Ustawienia',
-    'admin.mail.settings' => 'Ustawienia wysyłki wiadomości',
-    'admin.mail.template.list' => 'Szablony wiadomośći automatycznych',
-    'admin.update' => 'Aktualizacja',
+    
+    'boards' => 'Forum',
+    'get.settings' => 'Ustawienia',
+    'mail.settings' => 'Ustawienia wysyłki wiadomości',
+    'mail.template.list' => 'Szablony wiadomośći automatycznych',
+    'update' => 'Aktualizacja',
     'admin.users' => 'Użytkownicy',
-    'admin.user.edit' => 'Edycja użytkownika',
+    'user.edit' => 'Edycja użytkownika',
     'admin.groups' => 'Grupy użytkownków',
-    'admin.skinlist' => 'Skórki',
-    'admin.plugins.get' => 'Pluginy',
+    'skinlist' => 'Skórki',
+    'add.skin' => 'Dodaj skórkę',
+    'plugins.get' => 'Pluginy',
     'admin.pages' => 'Strony informacyjne',
+    'users.and.groups' => 'Użytkownicy i grupy',
+    'groups.add' => 'Dodaj grupę',
+    'groups.edit' => 'Edycja grupy',
+    'Module Position' => 'Pozycja modułu',
+    'modules.skin.edit' => 'Edycja modułu',
+    'modules.skin.get' => 'Moduły',
+    'modules.skin.new' => 'Dodaj moduł',
+    'page.edit' => 'Strona informacyjna',
+    'skin.twig.edit' => 'Pliki Twig',
+    'skin.css.edit' => 'Pliki CSS',
+    'skin.js.edit' => 'Pliki JavaScript',
+    'skin.addfile' => 'Dodaj plik',
+    
     
     'informations' => 'Informacje',
     'all plots' => 'Wszystkich tematów',
@@ -153,6 +173,11 @@ return [
     'menu links' => 'linki w menu',
     'advance url' => 'zaawansowane zarządzanie linkami',
     'url name' => 'nazwa linku',
-    'url' => 'link'
+    'url' => 'link',
     
+    'object' => 'Obiekt',
+    'date' => 'Data',
+    'logs' => 'Logi administratorów',
+    
+    'user.add' => 'Dodaj użytkownika',
 ];
