@@ -1,11 +1,23 @@
 <nav class="col-2 col-s-2 menu nav">
     <br />
 	<ul>
-      <li <?php if($step==0) echo'class="active"'; ?>>Start</li>
-	  <li <?php if($step==1) echo'class="active"'; ?>>Step 1</li>
-	  <li <?php if($step==2) echo'class="active"'; ?>>Step 2</li>
-	  <li <?php if($step==3) echo'class="active"'; ?>>Step 3</li>
-	  <li <?php if($step==4) echo'class="active"'; ?>>Step 4</li>
-	  <li <?php if($step==5) echo'class="active"'; ?>>Finish</li>
+      <li <?php if ($step==0) {
+    echo'class="active"';
+} ?>>Start</li>
+	  <li <?php if ($step==1) {
+    echo'class="active"';
+} ?>>Step 1</li>
+	  <li <?php if ($step==2) {
+    echo'class="active"';
+} ?>>Step 2</li>
+	  <li <?php if ($step==3) {
+    echo'class="active"';
+} ?>>Step 3</li>
+	  <li <?php if ($step==4) {
+    echo'class="active"';
+} ?>>Step 4</li>
+	  <li <?php if ($step==5) {
+    echo'class="active"';
+} ?>>Finish</li>
     </ul>
   </nav>
